Split Recall and Sample
