+++ ydb/library/workload/vector/ya.make
@@ -3,6 +3,7 @@ LIBRARY()
 SRCS(
     GLOBAL registrar.cpp
     vector_recall_evaluator.cpp
+    vector_sampler.cpp
     vector_sql.cpp
     vector_workload_generator.cpp
     vector_workload_params.cpp
